Error corrections for Gender Shades
Timnit Gebru <tgebru@gmail.com>
	
AttachmentsMon, Feb 27, 3:11 PM
	
to Sorelle, Christo, Joy
Hi friends,

I hope you've been well. I mentioned this to Sorelle in our last meeting but we found out that we need to make a correction to Gender Shades, after someone emailed us after trying to reproduce the numbers using the raw data listed in our supplementary materials.

We found out that we swapped true positive rates (TPR) and positive predictive value (PPV), on tables 4 and 5 of the paper and table 2 of the supplementary materials. This also changed the false positive rate (FPR) calculations. The corrections need to be made on pages 9 and 10 of the main paper and page 2 of the supplementary materials. We would like to make these corrections to the paper and supplementary materials. Nothing has changed about the conclusions about the paper or main results. We have received permission to update the paper from editors at the time, Sorelle Friedler and Christo Wilson. We received this permission on email on February 27, after we also sent the marked up version of the paper showing all the changes that will be made.